Avoids closing tweet stream when receiving data with no 'created_at'
We are sometimes receiving data without the 'created_at' key. This is probably
not a tweet, and we should look into it. This commit logs the error and the
data and keeps the stream from closing. Before this, the script was exiting
with status code 0, so supervisor wouldn't try to spawn a new process.