Uses a hash of the link to check for the articles existence
As of mongo 2.6 we cannot have index entries which are greater than 1Kb, and
some links are bigger than that. A sha1 has is guaranteed to have 20 bytes, so
we use that in order to be able to keep an index on the link which is at least
enough to check for it's existence.